requirement.txt
setuptools
wheel
langchain-text-splitters
langchain-community
langgraph
langchain-openai
langchain-core
sentence-transformers
pymilvus
pymilvus[model]
spacy[cuda12x]